rewored generate chunk into smart constructor and got rid of array of AnyRef
