+++ src/main/scala/org/terraframe/World.scala
@@ -1,67 +1,17 @@
 package org.terraframe
 
-import java.util.Random
 
 object World {
-  var x, y: Int = 0
-  var x2, y2, i, j, n, xpos, ypos, xpos2, ypos2: Int = _
-  var f: Double = _
-  var `type`: Short = _
 
-  var random: Random = _
+  import TerraFrame.mod
+  import TerraFrame.random
 
   var coordlist:Array2D[Boolean] = _
   var coordlist2:Array2D[Boolean] = _
 
-  def generateChunk(cx: Int, cy: Int, random: Random): Array[AnyRef] = {
-    val size: Int = TerraFrame.CHUNKBLOCKS
-    val blocks: Array3D[Int] = Array.ofDim(3,size,size)
-    val blockds: Array3D[Byte] = Array.ofDim(3,size,size)
-    val blockdns: Array2D[Byte] = Array.ofDim(size,size)
-    val blockbgs: Array2D[Byte] = Array.ofDim(size,size)
-    val blockts: Array2D[Byte] = Array.ofDim(size,size)
-    val lights: Array2D[Float] = Array.ofDim(size,size)
-    val power: Array3D[Float] = Array.ofDim(3,size,size)
-    val lsources: Array2D[Boolean] = Array.ofDim(size,size)
-    val zqn: Array2D[Byte] = Array.ofDim(size,size)
-    val pzqn: Array3D[Byte] = Array.ofDim(3,size,size)
-    val arbprd: Array3D[Boolean] = Array.ofDim(3,size,size)
-    val wcnct: Array2D[Boolean] = Array.ofDim(size,size)
-    val drawn: Array2D[Boolean] = Array.ofDim(size,size)
-    val rdrawn: Array2D[Boolean] = Array.ofDim(size,size)
-    val ldrawn: Array2D[Boolean] = Array.ofDim(size,size)
-    (0 until size).foreach { y =>
-      (0 until size).foreach { x =>
-        (0 until 3).foreach { l =>
-        if (l == 1 && cy*size+y >= PerlinNoise.perlinNoise((cx*size+x) / 10.0, 0.5, 0) * 30 + 50) {
-          blocks(l)(y)(x) = 1 // dirt
-        }
-        else {
-          blocks(l)(y)(x) = 0
-        }
-        arbprd(l)(y)(x) = false
-        power(l)(y)(x) = 0.toFloat
-      }
-        blockdns(y)(x) = random.nextInt(5).toByte
-        blockbgs(y)(x) = 0
-        blockts(y)(x) = random.nextInt(8).toByte
-        lights(y)(x) = 19.toFloat
-        lsources(y)(x) = false
-        wcnct(y)(x) = false
-        drawn(y)(x) = false
-        rdrawn(y)(x) = false
-        ldrawn(y)(x) = false
-        blockds(0)(y)(x) = 0.toByte
-        blockds(2)(y)(x) = 0.toByte
-      }
-    }
-    blockds(1) = generateOutlines(blocks(1))
-    val rv: Array[AnyRef] = Array(blocks, blockds, blockdns, blockbgs, blockts, lights, power, lsources, zqn, pzqn, arbprd, wcnct, drawn, rdrawn, ldrawn)
-    return rv
-  }
 
   def generateOutlines(blocks: Array2D[Int]): Array2D[Byte] = {
-    return generate2(blocks, false)
+    generate2(blocks, false)
   }
 
 
@@ -1343,11 +1293,6 @@ object World {
         }
       }
     }
-    return coordlist
-  }
-
-
-  def mod(a: Int, q: Int): Int = {
-    return TerraFrame.mod(a, q)
+    coordlist
   }
 }
